Balancing miniboss
