+++ combat/advancedCombat/index.js
@@ -147,6 +147,7 @@ const calculateCombatResult = async (progress) => {
 
 		const playerInfo = friendlyTeam.find((player) => player.account.userId === playerId);
 		// allows for multiple attack
+		console.log(playerInfo, "playerinfo");
 		const allowedNumOfAttacks = playerInfo.allowedNumOfAttacks || 1;
 		for (let i = 0; i < allowedNumOfAttacks; i += 1) {
 			const randomVictimInfo = opposingTeam[Math.floor(Math.random() * opposingTeam.length)];
