+++ game/_UNIVERSE/Misty Mountains/miniboss.js
@@ -4,18 +4,18 @@ module.exports = { "Graveward": {
 	stats:{
 		attack: 600,
 		health: 600,
-		difficulty: 35,
 	},
+	allowedNumOfAttacks: 2,
+	weapons: ["slash", "strike"],
 	rewards:{
 		dungeonKey: "Eridian Vase",
 		gold: 8888,
 		xp: 999,
 	},
-	rules: {
-		maxRounds: 2,
-		canKill:true,
-		allowArmy:false,
-		allowHelpers:true,
-		minRankToGetKey: 12,
+	minRankToGetKey: 12,
+	combatRules: {
+		maxRounds: 3,
+		armyAllowed: false,
+		helpersAllowed: true,
 	},
 }, };
